+++ usp-core-waap-header-filtering/step2/verify.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+# SPDX-FileCopyrightText: 2025 United Security Providers AG, Switzerland
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
+kubectl logs \
+  -n nextjs \
+  -l app.kubernetes.io/name=usp-core-waap \
+  --tail=-1 \
+  | grep "/api/hello"